Major changes
Fixed when it wouldn't parse with css links.

Better errors
    Changed it into an event emitter.
    Added massify.js
    bundle, and massify EventEmitters have the error events

Massify for fast bulk bundling
    Uses regular browserifies
    bundle.js now writes directly instead of the intermediate json file

Removed the intermediate file creation because it is probably too slow
It was neat, but pragmatics win

bulkcopy.js now does all the extra file, and folder copying

Don't know what else. I went crazy in the files because of so many changes I wanted.
Every thing seems to work anyway, and the cmd interface remains the same.